requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.2
flask==3.0.2
gunicorn==21.2.0